requirements.txt
tensorflow==2.2.0
scikit-learn~=0.24.1
pandas~=1.1.2
torch==1.6.0
pytorch-lightning==1.4.2
torchtext==0.7.0
pytorch-nlp==0.5.0
regex~=2020.11.13
ftfy~=6.0.3
loguru~=0.5.3
nltk==3.5
spacy==2.3.2
numpy~=1.19.2
utils~=1.0.1
sklearn~=0.0
fasttext
sqldf
captum